feat: rename from loader to muloader
